Пофикшены JacobySolver И SeidelSolver Внедрение в poisson.cpp

+++ ns/lib/slae/jacoby_solver.cpp
@@ -39,7 +39,7 @@ void JacobySolver::solve(const std::vector<double>& rhs, std::vector<double>& re
             double s = 0.0;
             for (int j = 1; j <= ind2 - ind1; j++){
 				int cc = c[j];
-				int vv = v[j];
+				double vv = v[j];
                 s -= vv * u_old[cc];
             }
             s += rhs[i];
